added perlin

+++ src/util/rgb.h
@@ -5,12 +5,12 @@ class cRGB
 {
 
 public:
-    float r;
-    float g;
-    float b;
+    double r;
+    double g;
+    double b;
 
-    cRGB() : r(0.f), g(0.f), b(0.f) {}
-    cRGB(float c0, float c1, float c2) : r(c0), g(c1), b(c2) {}
+    cRGB() : r(0.0), g(0.0), b(0.0) {}
+    cRGB(double c0, double c1, double c2) : r(c0), g(c1), b(c2) {}
 
     cRGB operator - () const { return cRGB(-r, -g, -b); }
 
@@ -32,11 +32,11 @@ public:
 
 };
 
-inline float LinearToGamma(float linear_component)
+inline double LinearToGamma(double linear_component)
 {
     if (linear_component > 0.f)
-        return std::sqrtf(linear_component);
-    return 0.f;
+        return std::sqrt(linear_component);
+    return 0.0;
 }
 
 inline std::ostream& operator << (std::ostream& out, const cRGB& rhs)
@@ -64,12 +64,12 @@ inline cRGB operator * (double t, const cRGB& rhs)
     return cRGB(t * rhs.r, t * rhs.g, t * rhs.b);
 }
 
-inline cRGB operator * (const cRGB lhs, float t)
+inline cRGB operator * (const cRGB lhs, double t)
 {
     return cRGB(lhs.r * t, lhs.g * t, lhs.b * t);
 }
 
-inline cRGB operator / (const cRGB& lhs, float t)
+inline cRGB operator / (const cRGB& lhs, double t)
 {
-    return (1.f / t) * lhs;
+    return (1.0 / t) * lhs;
 }
